+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ab-inbev/demo
 
+## 2.0.0
+
+### Major Changes
+
+- new tests
+
+### Patch Changes
+
+- Updated dependencies
+  - @schimufa/design-system@3.0.0
+
+## 1.0.0
+
+### Major Changes
+
+- New changes
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @schimufa/design-system@2.0.0
+
 ## 0.1.1
 
 ### Patch Changes